move handleFiles to middleware
- m-v0.2.2
